Created logging functionality, Implemented message logs in backup.py, changes in config.json
In order to use the logger (funtion to write log messages into a log file):
     "import this : from logger_settings import logger"
After that just call the function wherever you want like this:
     " logger.info("text") "
This function counts with this types of message : info, error, warning, critical and debug.

Moreover two parameters was removed from the json file: TEMP_DIR and LOG_DIR.
It was considered that this routes are static so the user should not be required to specify a path for these directories.